+++ src/tree_utils.h
@@ -49,5 +49,5 @@ string getNewickString (Tree * tree, string object);
 
 string double_to_str(double d);
 
-int get_num_possible_trees (int const& n, bool const& rooted);
+unsigned long int get_num_possible_trees (int const& n, bool const& rooted);
 #endif /* _TREE_UTILS_H_ */
